Polishing

+++ src/lib/components/Card.svelte
@@ -17,8 +17,11 @@
 
 	$: {
 		const query = new URLSearchParams($page.url.searchParams);
-		query.delete('container-preview');
-		query.append('container-preview', container.guid);
+		if (query.get('container-preview') === container.guid) {
+			query.delete('container-preview');
+		} else {
+			query.set('container-preview', container.guid);
+		}
 		containerPreviewURL = `?${query.toString()}`;
 	}
 
@@ -88,8 +91,8 @@
 				<Icon src={statusIcons.get(container.payload.status) ?? LightBulb} size="16" mini />
 				{$_(container.payload.status)}
 			</span>
-		{:else if 'topic' in container.payload}
-			<span class="badge">{$_(container.payload.topic)}</span>
+		{:else if 'strategyType' in container.payload}
+			<span class="badge">{$_(container.payload.strategyType)}</span>
 		{/if}
 		<a href={relatedToURL} class="button {relatedTo === container.guid ? 'is-active' : ''}">
 			<Icon src={Share} size="20" mini />
